backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
anthropic==0.84.0
python-dotenv==1.2.2
requests==2.32.5
python-multipart==0.0.22
google-generativeai
